+++ packages/koop-output-geoservices/CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* Error logging if code is 500
+* Replace message with Internal Server Error if 500 code
+
 ## [2.0.3] - 2020-10-19
 ### Fixed
 * Only normalize error if it exists.